Move log related assertions to fakes module
